Handle version lies Move wrapper code into implementation assembly Tweak test to ensure it fails if Nano changes behavior
